refactor(file-service): home the upload helpers with the code they belong to
Addresses review on #7763: ResourceUploadUtils was the wrong destination. The
two helpers have little in common with each other, and each has an existing
home that fits it better, so drop the new file and move them there.

- `put` -> LakeFSStorageClient. It is the middle step of the presigned
  multipart lifecycle that object already owns: initiate hands out the URLs,
  put uploads each part, and the ETag it returns is half of the
  (partNumber, eTag) pairs complete consumes. Placed next to those three so
  the lifecycle reads in order.

- `validateAndNormalizeFilePathOrThrow` -> ResourceNaming. It throws
  BadRequestException, so it belongs in the request layer rather than
  service/util, which holds operational helpers. It is the counterpart to
  validateName: the same "reject bad user input with a 400" contract applied
  to a path instead of a name.

DatasetResourcePathSpec tested neither of the things its name and package
claimed, so it becomes ResourceNamingSpec beside the code it exercises.

Pure move: both bodies are unchanged.

Part of #6494, groundwork for #6498.
